requirements.txt
flask
google-cloud-speech
google-cloud-texttospeech
python-dotenv
gunicorn
